ParkingApp/parking-reservation/requirements.txt
google-cloud-firestore
google-cloud-pubsub